Больше модулей (#11)
* CommonJS с ошибками

* CommonJS с ошибками

* #18 Больше модулей
