+++ src/js/main.js
@@ -1,1 +1,5 @@
 'use strict';
+
+require('../pictures');
+require('./resizer');
+require('./upload');
